Add files via upload
This is a small project I did to help me at my first job. We had to do data entry and for every data entry, it was required to write a 'diary'. This is a program for initial training. It guides the user what information is needed for a right entry and lastly, it prints out 5 of the practice names in a little report for my boss to see.